auto_remaster/sandbox/diffusers_flux2/requirements.txt
accelerate>=0.31.0
torchvision
transformers>=4.41.2
ftfy
tensorboard
Jinja2
peft>=0.11.1
sentencepiece
torchao
bitsandbytes
wandb
more_itertools